I have an algorithm that can find walls.
It identifies "blobs" of points that are close together, and
then finds the bounding box, which, if it is long and skinny
enough, qualifies it as a wall. This approach allowed the
computer to determine that my livingroom does indeed have
four walls.

Future steps:
- Use a least squares approximation to find a line of best fit
for each blob, and then count how many of the points in the
blob lie close to the line. This is the way it's done in
the RANSAC algorithm, and could replace the computationally
intensive bounding box method in use now.

- Eventually use this as part of a general-purpose SLAM
algorithm.